requirements.txt
aiogram==3.15.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
requests==2.32.3
aspose-words==25.9.0
docx2txt==0.8